+++ components/netsuite/src/types/index.ts
@@ -2,6 +2,10 @@ export type {
   NetSuitePollingState,
   PollingChangesResult,
 } from "./PollingState";
-export type { PollingTriggerObject } from "./PollingTriggerObject";
+export type {
+  PollingChangesObject,
+  PollingRecordChange,
+  PollingTriggerObject,
+} from "./PollingTriggerObject";
 export type { FetchAllRecordsParams, Record } from "./Record";
 export type { RecordType } from "./RecordType";