+++ components/netsuite/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./auth";
+export * from "./cleanFunctions";
 export * from "./polling";
 export * from "./records";